requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
python-multipart>=0.0.9

tensorflow==2.15.0
numpy>=1.23,<2.0
ultralytics>=8.0

pillow>=10.0
opencv-python-headless>=4.8